Move migrate_key script to docs/ dir
Script does not have to be part of the public API/CLI or be packaged.
Just want to have it in the repo and use existing CI.
